# Partially undo my optimization of migration_cache.Writer