`io.Writer` says that
> Implementations must not retain p (the buffer passed to Write)

The `asyncWriter` I introduced definitely violates this. In practice, this worked for us, because the buffer comes from proto parsing code, which doesn't reuse the buffer. Still, I don't feel comfortable keeping this foot-gun around.

I've change the implementation to be pretty similar to how it was before my optimization, but it uses a single `sync.WaitGroup`, instead of creating a new `errgroup.Group` for every call. Performance wise, this new version ends up being roughly half way between the original version and the `asyncWriter`.

It's also a lot simpler.

Benchmark results:
```
                                 │  original   │         broken-optimization         │          new-optimization          │
                                 │   sec/op    │   sec/op     vs base                │   sec/op     vs base               │
Set/LocalMigration10/AC-24         35.25µ ± 2%   32.00µ ± 4%   -9.24% (p=0.000 n=11)   33.43µ ± 3%  -5.17% (p=0.000 n=11)
Set/LocalMigration10/CAS-24        29.87µ ± 2%   27.02µ ± 6%   -9.54% (p=0.000 n=11)   27.92µ ± 1%  -6.52% (p=0.000 n=11)
Set/LocalMigration100/AC-24        44.68µ ± 1%   40.94µ ± 3%   -8.38% (p=0.000 n=11)   42.62µ ± 2%  -4.61% (p=0.000 n=11)
Set/LocalMigration100/CAS-24       41.13µ ± 2%   37.64µ ± 1%   -8.48% (p=0.000 n=11)   38.65µ ± 2%  -6.03% (p=0.000 n=11)
Set/LocalMigration1000/AC-24       67.21µ ± 5%   59.44µ ± 6%  -11.56% (p=0.000 n=11)   64.99µ ± 2%  -3.30% (p=0.001 n=11)
Set/LocalMigration1000/CAS-24      61.30µ ± 4%   54.04µ ± 6%  -11.83% (p=0.005 n=11)   58.57µ ± 4%  -4.46% (p=0.000 n=11)
Set/LocalMigration10000/AC-24      281.7µ ± 1%   254.8µ ± 2%   -9.56% (p=0.000 n=11)   281.0µ ± 1%       ~ (p=0.508 n=11)
Set/LocalMigration10000/CAS-24     280.4µ ± 1%   250.4µ ± 1%  -10.69% (p=0.000 n=11)   282.1µ ± 1%       ~ (p=0.401 n=11)
Set/LocalMigration1000000/AC-24    3.598m ± 1%   3.533m ± 2%   -1.79% (p=0.005 n=11)   3.570m ± 2%       ~ (p=0.478 n=11)
Set/LocalMigration1000000/CAS-24   3.623m ± 6%   3.558m ± 3%   -1.79% (p=0.023 n=11)   3.597m ± 1%       ~ (p=0.270 n=11)
Set/DistMigration10/AC-24          150.4µ ± 3%   142.5µ ± 3%   -5.22% (p=0.001 n=11)   146.9µ ± 1%  -2.31% (p=0.003 n=11)
Set/DistMigration10/CAS-24         98.19µ ± 2%   94.43µ ± 4%        ~ (p=0.065 n=11)   97.83µ ± 1%       ~ (p=0.847 n=11)
Set/DistMigration100/AC-24         165.6µ ± 1%   156.6µ ± 3%   -5.43% (p=0.001 n=11)   162.1µ ± 1%  -2.10% (p=0.000 n=11)
Set/DistMigration100/CAS-24        99.43µ ± 2%   97.78µ ± 3%        ~ (p=0.243 n=11)   99.18µ ± 2%       ~ (p=0.606 n=11)
Set/DistMigration1000/AC-24        190.6µ ± 1%   186.8µ ± 2%   -2.01% (p=0.000 n=11)   188.2µ ± 1%  -1.26% (p=0.000 n=11)
Set/DistMigration1000/CAS-24       100.7µ ± 2%   105.8µ ± 4%   +5.04% (p=0.001 n=11)   100.7µ ± 1%       ~ (p=0.748 n=11)
Set/DistMigration10000/AC-24       442.2µ ± 1%   391.0µ ± 3%  -11.56% (p=0.000 n=11)   440.4µ ± 0%  -0.39% (p=0.047 n=11)
Set/DistMigration10000/CAS-24      109.8µ ± 1%   107.7µ ± 2%   -1.93% (p=0.040 n=11)   108.8µ ± 3%  -0.90% (p=0.016 n=11)
Set/DistMigration1000000/AC-24     4.480m ± 1%   4.441m ± 2%        ~ (p=0.101 n=11)   4.469m ± 1%       ~ (p=0.193 n=11)
Set/DistMigration1000000/CAS-24    1.188m ± 2%   1.207m ± 3%        ~ (p=0.151 n=11)   1.187m ± 1%       ~ (p=1.000 n=11)
geomean                            199.4µ        188.3µ        -5.56%                  195.5µ       -1.98%

                                 │   original    │          broken-optimization           │           new-optimization            │
                                 │      B/s      │      B/s        vs base                │     B/s       vs base                 │
Set/LocalMigration10/AC-24         273.4Ki ±  4%    302.7Ki ±  3%  +10.71% (p=0.000 n=11)   293.0Ki ± 3%  +7.14% (p=0.000 n=11)
Set/LocalMigration10/CAS-24        322.3Ki ±  3%    361.3Ki ±  5%  +12.12% (p=0.000 n=11)   351.6Ki ± 0%  +9.09% (p=0.000 n=11)
Set/LocalMigration100/AC-24        2.136Mi ±  2%    2.327Mi ±  4%   +8.93% (p=0.000 n=11)   2.241Mi ± 1%  +4.91% (p=0.000 n=11)
Set/LocalMigration100/CAS-24       2.317Mi ±  2%    2.537Mi ±  3%   +9.47% (p=0.000 n=11)   2.470Mi ± 2%  +6.58% (p=0.000 n=11)
Set/LocalMigration1000/AC-24       14.19Mi ±  6%    16.04Mi ±  6%  +13.04% (p=0.000 n=11)   14.68Mi ± 2%  +3.43% (p=0.001 n=11)
Set/LocalMigration1000/CAS-24      15.55Mi ±  3%    17.64Mi ± 14%  +13.43% (p=0.005 n=11)   16.29Mi ± 2%  +4.72% (p=0.000 n=11)
Set/LocalMigration10000/AC-24      33.86Mi ±  2%    37.43Mi ±  3%  +10.56% (p=0.000 n=11)   33.94Mi ± 1%       ~ (p=0.508 n=11)
Set/LocalMigration10000/CAS-24     34.02Mi ±  3%    38.09Mi ±  1%  +11.97% (p=0.000 n=11)   33.80Mi ± 2%       ~ (p=0.374 n=11)
Set/LocalMigration1000000/AC-24    265.1Mi ±  1%    269.9Mi ±  2%   +1.82% (p=0.005 n=11)   267.1Mi ± 2%       ~ (p=0.467 n=11)
Set/LocalMigration1000000/CAS-24   263.2Mi ± 13%    268.0Mi ±  3%   +1.82% (p=0.023 n=11)   265.2Mi ± 1%       ~ (p=0.270 n=11)
Set/DistMigration10/AC-24          68.36Ki ± 14%    68.36Ki ±  0%        ~ (p=0.586 n=11)   68.36Ki ± 0%       ~ (p=0.586 n=11)
Set/DistMigration10/CAS-24         97.66Ki ±  0%   107.42Ki ± 18%        ~ (p=0.063 n=11)   97.66Ki ± 0%       ~ (p=1.000 n=11) ¹
Set/DistMigration100/AC-24         585.9Ki ±  2%    625.0Ki ±  6%   +6.67% (p=0.001 n=11)   605.5Ki ± 2%  +3.33% (p=0.000 n=11)
Set/DistMigration100/CAS-24        986.3Ki ±  2%    996.1Ki ±  4%        ~ (p=0.221 n=11)   986.3Ki ± 1%       ~ (p=0.726 n=11)
Set/DistMigration1000/AC-24        5.007Mi ±  1%    5.102Mi ±  2%   +1.90% (p=0.000 n=11)   5.064Mi ± 1%  +1.14% (p=0.000 n=11)
Set/DistMigration1000/CAS-24       9.470Mi ±  1%    9.012Mi ±  4%   -4.83% (p=0.001 n=11)   9.470Mi ± 1%       ~ (p=0.759 n=11)
Set/DistMigration10000/AC-24       21.57Mi ±  1%    24.39Mi ±  3%  +13.04% (p=0.000 n=11)   21.65Mi ± 0%       ~ (p=0.058 n=11)
Set/DistMigration10000/CAS-24      86.83Mi ±  2%    88.54Mi ±  5%   +1.97% (p=0.040 n=11)   87.62Mi ± 2%  +0.91% (p=0.016 n=11)
Set/DistMigration1000000/AC-24     212.9Mi ±  1%    214.7Mi ±  2%        ~ (p=0.091 n=11)   213.4Mi ± 1%       ~ (p=0.193 n=11)
Set/DistMigration1000000/CAS-24    802.7Mi ±  2%    790.0Mi ±  2%        ~ (p=0.151 n=11)   803.2Mi ± 2%       ~ (p=1.000 n=11)
geomean                            7.581Mi          8.035Mi         +5.99%                  7.741Mi       +2.11%
¹ all samples are equal

                                 │   original    │          broken-optimization          │           new-optimization           │
                                 │     B/op      │     B/op       vs base                │     B/op       vs base               │
Set/LocalMigration10/AC-24         15.17Ki ±  0%   15.30Ki ±  0%   +0.89% (p=0.001 n=11)   14.83Ki ±  0%  -2.21% (p=0.000 n=11)
Set/LocalMigration10/CAS-24        11.16Ki ±  0%   11.30Ki ±  0%   +1.19% (p=0.000 n=11)   10.83Ki ±  0%  -3.01% (p=0.000 n=11)
Set/LocalMigration100/AC-24        15.41Ki ±  0%   15.55Ki ±  0%   +0.89% (p=0.000 n=11)   15.08Ki ±  0%  -2.14% (p=0.000 n=11)
Set/LocalMigration100/CAS-24       11.41Ki ±  0%   11.54Ki ±  0%   +1.16% (p=0.000 n=11)   11.07Ki ±  0%  -2.97% (p=0.000 n=11)
Set/LocalMigration1000/AC-24       18.49Ki ±  0%   18.62Ki ±  0%   +0.71% (p=0.000 n=11)   18.15Ki ±  0%  -1.83% (p=0.000 n=11)
Set/LocalMigration1000/CAS-24      14.47Ki ±  0%   14.60Ki ±  0%   +0.91% (p=0.000 n=11)   14.15Ki ±  0%  -2.21% (p=0.000 n=11)
Set/LocalMigration10000/AC-24      20.60Ki ±  0%   20.72Ki ±  0%   +0.59% (p=0.000 n=11)   20.25Ki ±  0%  -1.68% (p=0.000 n=11)
Set/LocalMigration10000/CAS-24     16.57Ki ±  0%   16.70Ki ±  0%   +0.75% (p=0.001 n=11)   16.24Ki ±  0%  -2.03% (p=0.000 n=11)
Set/LocalMigration1000000/AC-24    30.45Ki ± 30%   27.04Ki ± 22%        ~ (p=0.438 n=11)   27.06Ki ± 12%       ~ (p=0.217 n=11)
Set/LocalMigration1000000/CAS-24   27.06Ki ± 31%   20.80Ki ± 16%  -23.13% (p=0.013 n=11)   26.42Ki ± 24%       ~ (p=0.217 n=11)
Set/DistMigration10/AC-24          44.30Ki ±  0%   44.27Ki ±  0%   -0.08% (p=0.040 n=11)   43.84Ki ±  0%  -1.05% (p=0.000 n=11)
Set/DistMigration10/CAS-24         31.83Ki ±  0%   31.82Ki ±  0%        ~ (p=0.197 n=11)   31.83Ki ±  0%       ~ (p=0.595 n=11)
Set/DistMigration100/AC-24         44.81Ki ±  0%   44.78Ki ±  0%   -0.08% (p=0.000 n=11)   44.36Ki ±  0%  -1.00% (p=0.000 n=11)
Set/DistMigration100/CAS-24        32.26Ki ±  0%   32.26Ki ±  0%        ~ (p=0.053 n=11)   32.26Ki ±  0%  -0.01% (p=0.015 n=11)
Set/DistMigration1000/AC-24        48.18Ki ±  0%   48.11Ki ±  0%   -0.14% (p=0.001 n=11)   47.72Ki ±  0%  -0.95% (p=0.000 n=11)
Set/DistMigration1000/CAS-24       33.22Ki ±  0%   33.22Ki ±  0%        ~ (p=0.508 n=11)   33.22Ki ±  0%       ~ (p=0.175 n=11)
Set/DistMigration10000/AC-24       59.81Ki ±  0%   59.70Ki ±  0%   -0.18% (p=0.001 n=11)   59.34Ki ±  0%  -0.79% (p=0.000 n=11)
Set/DistMigration10000/CAS-24      41.71Ki ±  0%   41.70Ki ±  0%   -0.03% (p=0.020 n=11)   41.70Ki ±  0%       ~ (p=0.249 n=11)
Set/DistMigration1000000/AC-24     1.074Mi ±  2%   1.076Mi ±  2%        ~ (p=0.652 n=11)   1.069Mi ±  2%       ~ (p=0.797 n=11)
Set/DistMigration1000000/CAS-24    1.032Mi ±  0%   1.030Mi ±  0%        ~ (p=0.300 n=11)   1.032Mi ±  1%       ~ (p=0.949 n=11)
geomean                            36.84Ki         36.26Ki         -1.57%                  36.17Ki        -1.83%

                                 │  original  │         broken-optimization         │          new-optimization           │
                                 │ allocs/op  │ allocs/op   vs base                 │ allocs/op   vs base                 │
Set/LocalMigration10/AC-24         232.0 ± 0%   228.0 ± 0%  -1.72% (p=0.000 n=11)     226.0 ± 0%  -2.59% (p=0.000 n=11)
Set/LocalMigration10/CAS-24        178.0 ± 0%   174.0 ± 0%  -2.25% (p=0.000 n=11)     172.0 ± 0%  -3.37% (p=0.000 n=11)
Set/LocalMigration100/AC-24        228.0 ± 0%   224.0 ± 0%  -1.75% (p=0.000 n=11)     222.0 ± 0%  -2.63% (p=0.000 n=11)
Set/LocalMigration100/CAS-24       174.0 ± 0%   170.0 ± 0%  -2.30% (p=0.000 n=11)     168.0 ± 0%  -3.45% (p=0.000 n=11)
Set/LocalMigration1000/AC-24       228.0 ± 0%   224.0 ± 0%  -1.75% (p=0.000 n=11)     222.0 ± 0%  -2.63% (p=0.000 n=11)
Set/LocalMigration1000/CAS-24      174.0 ± 0%   170.0 ± 0%  -2.30% (p=0.000 n=11)     168.0 ± 0%  -3.45% (p=0.000 n=11)
Set/LocalMigration10000/AC-24      283.0 ± 0%   279.0 ± 0%  -1.41% (p=0.000 n=11)     277.0 ± 0%  -2.12% (p=0.000 n=11)
Set/LocalMigration10000/CAS-24     229.0 ± 0%   225.0 ± 0%  -1.75% (p=0.000 n=11)     223.0 ± 0%  -2.62% (p=0.000 n=11)
Set/LocalMigration1000000/AC-24    290.0 ± 1%   287.0 ± 1%  -1.03% (p=0.000 n=11)     284.0 ± 1%  -2.07% (p=0.000 n=11)
Set/LocalMigration1000000/CAS-24   240.0 ± 0%   236.0 ± 1%  -1.67% (p=0.000 n=11)     235.0 ± 1%  -2.08% (p=0.000 n=11)
Set/DistMigration10/AC-24          691.0 ± 0%   684.0 ± 0%  -1.01% (p=0.000 n=11)     683.0 ± 0%  -1.16% (p=0.000 n=11)
Set/DistMigration10/CAS-24         495.0 ± 0%   495.0 ± 0%       ~ (p=0.311 n=11)     495.0 ± 0%       ~ (p=1.000 n=11)
Set/DistMigration100/AC-24         687.0 ± 0%   679.0 ± 0%  -1.16% (p=0.000 n=11)     679.0 ± 0%  -1.16% (p=0.000 n=11)
Set/DistMigration100/CAS-24        497.0 ± 0%   496.0 ± 0%  -0.20% (p=0.008 n=11)     496.0 ± 0%  -0.20% (p=0.002 n=11)
Set/DistMigration1000/AC-24        682.0 ± 0%   674.0 ± 0%  -1.17% (p=0.000 n=11)     674.0 ± 0%  -1.17% (p=0.000 n=11)
Set/DistMigration1000/CAS-24       492.0 ± 0%   492.0 ± 0%       ~ (p=1.000 n=11)     492.0 ± 0%       ~ (p=1.000 n=11) ¹
Set/DistMigration10000/AC-24       743.0 ± 0%   735.0 ± 0%  -1.08% (p=0.000 n=11)     735.0 ± 0%  -1.08% (p=0.000 n=11)
Set/DistMigration10000/CAS-24      493.0 ± 0%   493.0 ± 0%       ~ (p=1.000 n=11) ¹   493.0 ± 0%       ~ (p=1.000 n=11) ¹
Set/DistMigration1000000/AC-24     822.0 ± 0%   822.0 ± 0%       ~ (p=0.635 n=11)     817.0 ± 0%  -0.61% (p=0.013 n=11)
Set/DistMigration1000000/CAS-24    608.0 ± 0%   610.0 ± 0%  +0.33% (p=0.001 n=11)     610.0 ± 0%       ~ (p=0.129 n=11)
geomean                            368.3        364.1       -1.12%                    362.3       -1.61%
```